Migrate password manager to typescript (partially)
