+++ microservices/password-manager/app.ts
@@ -28,14 +28,9 @@ const logger = winston.createLogger({
   transports: [new winston.transports.Console()],
 });
 
-app.get("/", (req, res) => {
-  res.status(200).send({
-    status: "Hello world",
-  });
-});
-
 PrimistoreController(app, logger);
 
-app.listen(process.env.PORT || 4000, process.env.HOST || "127.0.0.1", () => {
-  createDirectoriesIfNotExistSync(PRIMISTORE_DIR);
+const port = process.env.PORT ? parseInt(process.env.PORT) : 4000;
+app.listen(port, process.env.HOST || "127.0.0.1", () => {
+    createDirectoriesIfNotExistSync(PRIMISTORE_DIR);
 });
