+++ microservices/password-manager/package.json
@@ -6,13 +6,16 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node app.js",
-    "dev": "nodemon app.js"
+    "start": "tsc && node dist/app.js",
+    "dev": "tsc && nodemon dist/app.js"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/node": "^20.12.7",
     "cors": "^2.8.5",
     "express": "^4.18.2",
     "mongoose": "^8.0.0",
