+++ microservices/password-manager/utils/date-utils.ts
@@ -1,9 +1,8 @@
-const getCurrentTime = () => {
+const getCurrentTime = (): string => {
   const currentDate = new Date();
-
   const systemTimeZone = Intl.DateTimeFormat().resolvedOptions().timeZone;
 
-  const options = {
+  const options: Intl.DateTimeFormatOptions = {
     year: "numeric",
     month: "2-digit",
     day: "2-digit",
